Move stat collection back before choosing int candidates (duh) and fix an overflow when more instructions are eliminated than the total (clearly a bug in itself, but fairly harmless)

+++ llvm/lib/Analysis/Integrator/IntBenefit.cpp
@@ -58,7 +58,7 @@ int64_t IntegrationAttempt::getResidualInstructions() {
     return residualInstructions;
 
   // Total excluding explored child loops:
-  int64_t totalInstructions = getTotalInstructions() - getElimdInstructions();
+  int64_t totalInstructions = ((int)getTotalInstructions()) - ((int)getElimdInstructions());
 
   for(DenseMap<const Loop*, PeelAttempt*>::iterator it = peelChildren.begin(), it2 = peelChildren.end(); it != it2; ++it) {
